+++ test/fixtures/custom-project/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
-## v1.2.3
+## v1.2.3 — 2025-04-06
 
 - Fixture release.